@@ -62,8 +62,8 @@
     "oskari-rpc": "^2.1.0",
     "parse-css-color": "^0.2.1",
     "process": "^0.11.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-draft-wysiwyg": "^1.15.0",
     "react-dropzone": "^11.7.1",
     "react-markdown": "^7.1.2",
@@ -76,24 +76,23 @@
   "devDependencies": {
     "@babel/preset-env": "^7.27.2",
     "@babel/preset-react": "^7.27.1",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/react": "^16.3.0",
     "@types/draft-js": "^0.11.18",
-    "@types/enzyme": "^3.10.19",
     "@types/file-saver": "^2.0.7",
     "@types/geojson": "^7946.0.16",
     "@types/jest": "^28.1.8",
     "@types/markdown-draft-js": "^2.2.7",
     "@types/node": "^20.19.1",
-    "@types/react": "^17.0.87",
-    "@types/react-dom": "^17.0.26",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "@types/react-draft-wysiwyg": "^1.13.8",
     "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^8.34.1",
     "@typescript-eslint/parser": "^8.34.1",
     "@vitejs/plugin-react": "^4.6.0",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
     "babel-eslint": "^10.1.0",
     "dotenv": "^10.0.0",
-    "enzyme": "^3.11.0",
     "eslint": "^9.29.0",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-node": "^11.1.0",
