@@ -47,7 +47,7 @@ async function start() {
           logger.info(message.split('\n')[0]);
         },
       },
-    })
+    }),
   );
 
   if (process.env.AUTH_ENABLED === 'true') {
@@ -64,7 +64,7 @@ async function start() {
     ensureAuthenticated({
       redirectToLogin: true,
     }),
-    express.static('static/admin')
+    express.static('static/admin'),
   );
   app.use('/', express.static('static'));
 
@@ -79,16 +79,16 @@ async function start() {
     }),
     (req, res) => {
       res.sendFile(path.join(__dirname, '../static/admin/index.html'));
-    }
+    },
   );
 
   // Serve frontend files from remaining URLs
-  app.get('/*', (req, res, next) => {
+  app.get('/*', (req, res, _next) => {
     res.sendFile(path.join(__dirname, '../static/index.html'));
   });
 
   // Default error handler
-  app.use((error: HttpResponseError, req, res, next) => {
+  app.use((error: HttpResponseError, req, res, _next) => {
     logger.error(`Request error: ${error.message}`);
     console.error(error);
     res.status(error.status || 500);
